tcpconnect fixes lots of hexen2 fixes fixed clipped decals again, still not using any... fixed zips over 2g rewrote bloom to use glsl. should be slightly more usable now. lots more hexen2 fixes
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@3957 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/unzip.c
@@ -209,7 +209,8 @@ local unsigned long unzlocal_SearchCentralDir(vfsfile_t *fin) {
 		if (!VFS_SEEK(fin, uReadPos))
 			break;
 
-		if (VFS_READ(fin,buf,(unsigned int)uReadSize)!=uReadSize) break;
+		if (VFS_READ(fin,buf,uReadSize)!=uReadSize)
+			break;
 
 		for (i=(int)uReadSize-3; (i--)>0;)
 			if (((*(buf+i))==0x50) && ((*(buf+i+1))==0x4b) && 
@@ -218,7 +219,8 @@ local unsigned long unzlocal_SearchCentralDir(vfsfile_t *fin) {
 				break;
 			}
 
-		if (uPosFound!=0) break;
+		if (uPosFound!=0)
+			break;
 	}
 	TRYFREE(buf);
 	return uPosFound;
